Disable all tests and documentation
This PR disables, during the building of parts, all the tests
and documentation build, and also enables building gobject
introspection data and VAPI files, to ensure that all libraries
are available in any language.

It also removes the usr/share/gtk-doc folder, since it's not
needed.

All this saves a little bit of space and slightly reduces the
build time.