Fix dependencies for gobject-introspection and Vapigen
Some parts use gobject-introspection or vapigen, which require
Vala and Gobject-introspection parts being built before.
Unfortunately, some dependencies didn't account for this.

Also, the PATH environment was wrong, and didn't prioritize the
STAGE, so in some cases the wrong binary was being used.

Due to all of this, building libportal failed when using
gobject-introspection greater than 1.82.0.

This PR fixes all these things, and also adds an extra check in
check_dependencies.py to ensure that any part that is building
gobject-introspection or vapi files have the correct
dependencies in the `after` entry.