debs: Install all the build dependencies in the SDK
The GNOME SDK should contain all the required dev-dependencies of the
libraries we ship, including the core files dependencies.

Now, even though core includes already some of the runtime libraries it
does not include their *.so counterparts, and then that's now our duty
to provide them all and in a working state.

Since we can't really symlink to external paths in a snap, we should
provide all the libraries here, since there's no way to fix the *.so files
that are used at link time.

Dropping them would be wrong though, since there may be libraries in the
SDK that requires them at link time only, and it's not up to final
applications to manage that.