Add missing -dev files (#313)
* Add missing -dev files

Some packages are added to the stage, but not their corresponding
development packages. This patch fixes this.

* Sort packages and remove unneeded

The unneeded packages are those that already install the -dev,
so the "normal" one should not be needed.

* Deb packages required

Some -dev packages seem to not force installation of the
corresponding library package, so they must be selected manually.

* Include comment about why runtime pkgs are added